+++ examples/use-cases/cloud_router_aws_google/cloud_router_connection_aws.tf
@@ -61,12 +61,10 @@ resource "packetfabric_cloud_router_bgp_session" "crbs_1" {
   prefixes {
     prefix = var.gcp_subnet_cidr1
     type   = "out" # Allowed Prefixes to Cloud
-    order  = 0
   }
   prefixes {
     prefix = var.aws_vpc_cidr1
     type   = "in" # Allowed Prefixes from Cloud
-    order  = 0
   }
 }
 output "packetfabric_cloud_router_bgp_session_crbs_1" {