+++ examples/use-cases/cloud_router_google_azure/provider.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     packetfabric = {
       source  = "PacketFabric/packetfabric"
-      version = ">= 1.0.4"
+      version = ">= 1.1.0"
     }
     google = {
       source  = "hashicorp/google"