+++ examples/use-cases/cloud_router_google_ipsec/cloud_router_connection_google.tf
@@ -127,12 +127,10 @@ resource "packetfabric_cloud_router_bgp_session" "crbs_1" {
   prefixes {
     prefix = var.ipsec_subnet_cidr2
     type   = "out" # Allowed Prefixes to Cloud
-    order  = 0
   }
   prefixes {
     prefix = var.google_subnet_cidr1
     type   = "in" # Allowed Prefixes from Cloud
-    order  = 0
   }
 
   # # workaround until we can use lifecycle into Terraform gcloud Module