+++ examples/use-cases/cloud_router_ibm_oracle/cloud_router_connection_ibm.tf
@@ -37,12 +37,10 @@ resource "packetfabric_cloud_router_bgp_session" "crbs_1" {
   prefixes {
     prefix = var.oracle_subnet_cidr1
     type   = "out" # Allowed Prefixes to Cloud
-    order  = 0
   }
   prefixes {
     prefix = var.ibm_vpc_cidr1
     type   = "in" # Allowed Prefixes from Cloud
-    order  = 0
   }
 }
 output "packetfabric_cloud_router_bgp_session_crbs_1" {