+++ internal/provider/resource_port.go
@@ -39,7 +39,6 @@ func resourceInterfaces() *schema.Resource {
 			"autoneg": {
 				Type:        schema.TypeBool,
 				Optional:    true,
-				Default:     false,
 				Description: "Only applicable to 1Gbps ports. Controls whether auto negotiation is on (true) or off (false). The request will fail if specified with 10Gbps. ",
 			},
 			"description": {
@@ -165,17 +164,13 @@ func resourceDeleteInterface(ctx context.Context, d *schema.ResourceData, m inte
 	c := m.(*packetfabric.PFClient)
 	c.Ctx = ctx
 	var diags diag.Diagnostics
-	resp, err := c.DeletePort(d.Id())
+	_, err := c.DeletePort(d.Id())
 	time.Sleep(30 * time.Second)
 	if err != nil {
 		return diag.FromErr(err)
 	}
 	d.SetId("")
-	return append(diags, diag.Diagnostic{
-		Severity: diag.Warning,
-		Summary:  "Port Delete",
-		Detail:   resp.Message,
-	})
+	return diags
 }
 
 func extractInterface(d *schema.ResourceData) packetfabric.Interface {