+++ internal/provider/resource_third_party_virtual_circuit.go
@@ -78,7 +78,7 @@ func resourceThirdPartyVirtualCircuit() *schema.Resource {
 							Type:         schema.TypeString,
 							Optional:     true,
 							ForceNew:     true,
-							ValidateFunc: validation.StringInSlice(ixVcSpeedOptions(), true),
+							ValidateFunc: validation.StringInSlice(speedOptions(), true),
 							Description:  "The desired speed of the new connection. Only applicable if `longhaul_type` is \"dedicated\" or \"hourly\".\n\n\tEnum: [\"50Mbps\" \"100Mbps\" \"200Mbps\" \"300Mbps\" \"400Mbps\" \"500Mbps\" \"1Gbps\" \"2Gbps\" \"5Gbps\" \"10Gbps\" \"20Gbps\" \"30Gbps\" \"40Gbps\" \"50Gbps\" \"60Gbps\" \"80Gbps\" \"100Gbps\"]",
 						},
 						"subscription_term": {
@@ -179,15 +179,8 @@ func resourceThirdPartyVirtualCircuitDelete(ctx context.Context, d *schema.Resou
 	c := m.(*packetfabric.PFClient)
 	c.Ctx = ctx
 	var diags diag.Diagnostics
-	if msg, err := c.DeleteHostedMktConnection(d.Id()); err != nil {
+	if _, err := c.DeleteHostedMktConnection(d.Id()); err != nil {
 		return diag.FromErr(err)
-	} else {
-		diags = make(diag.Diagnostics, 0)
-		diags = append(diags, diag.Diagnostic{
-			Severity: diag.Warning,
-			Summary:  "Third Party VC delete result",
-			Detail:   msg,
-		})
 	}
 	d.SetId("")
 	return diags